+++ plugin/pxr/maya/lib/usdMaya/CMakeLists.txt
@@ -80,6 +80,7 @@ pxr_test_scripts(
         testenv/testUsdMayaAdaptor.py
         testenv/testUsdMayaAdaptorGeom.py
         testenv/testUsdMayaAdaptorMetadata.py
+        testenv/testUsdMayaAdaptorUndoRedo.py
         testenv/testUsdMayaBlockSceneModificationContext.py
         testenv/testUsdMayaDiagnosticDelegate.py
         testenv/testUsdMayaGetVariantSetSelections.py
@@ -276,6 +277,18 @@ pxr_register_test(testUsdMayaAdaptorMetadata
     CUSTOM_PYTHON ${MAYA_PY_EXECUTABLE}
     COMMAND "${TEST_INSTALL_PREFIX}/tests/testUsdMayaAdaptorMetadata"
     TESTENV testUsdMayaAdaptorMetadata
+    ENV
+        MAYA_PLUG_IN_PATH=${TEST_INSTALL_PREFIX}/maya/plugin
+        MAYA_SCRIPT_PATH=${TEST_INSTALL_PREFIX}/maya/lib/usd/usdMaya/resources
+        MAYA_DISABLE_CIP=1
+        MAYA_NO_STANDALONE_ATEXIT=1
+        MAYA_APP_DIR=<PXR_TEST_DIR>/maya_profile
+)
+
+pxr_register_test(testUsdMayaAdaptorUndoRedo
+    CUSTOM_PYTHON ${MAYA_PY_EXECUTABLE}
+    COMMAND "${TEST_INSTALL_PREFIX}/tests/testUsdMayaAdaptorUndoRedo"
+    TESTENV testUsdMayaAdaptorUndoRedo
     ENV
         MAYA_PLUG_IN_PATH=${TEST_INSTALL_PREFIX}/maya/plugin
         MAYA_SCRIPT_PATH=${TEST_INSTALL_PREFIX}/maya/lib/usd/usdMaya/resources
